requirements.txt
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.4
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==3.0.4
